+++ app/controllers/publisher.controller.js
@@ -2,6 +2,18 @@
 const fieldList = ["MaNXB", "TenNXB", "DiaChi"]
 const collection = "NhaXuatBan"
 const singleCollectionName = "NhaXuatBan"
+const sachFieldList = [
+  "MaSach",
+  "TenSach",
+  "DonGia",
+  "SoQuyen",
+  "DangYeuCau",
+  "DangMuon",
+  "NamXuatBan",
+  "MaNXB",
+  "NguonGoc/TacGia",
+]
+const sachCollection = "Sach"
 //configure in create, update method also
 //
 
@@ -133,10 +145,20 @@ exports.update = async (req, res, next) => {
   }
 }
 exports.delete = async (req, res, next) => {
+  const id = req.params.id
   try {
     const dbService = new DatabaseService(MongoDB.client, collection, fieldList)
 
-    let document = await dbService.delete(req.params.id)
+    if (checkUsed(id)) {
+      return res.send(
+        ResponseTemplate(
+          false,
+          "Không thể xóa! Nhà xuất bản đang được sử dụng trong hệ thống!",
+          null
+        )
+      )
+    }
+    let document = await dbService.delete(id)
 
     if (!document) {
       return next(new ApiError(404, `${singleCollectionName} not found`))
@@ -152,7 +174,7 @@ exports.delete = async (req, res, next) => {
     return next(
       new ApiError(
         500,
-        `Could not delete ${singleCollectionName} with id=${req.params.id}`
+        `Could not delete ${singleCollectionName} with id=${id}`
       )
     )
   }
@@ -160,7 +182,15 @@ exports.delete = async (req, res, next) => {
 exports.deleteAll = async (req, res, next) => {
   try {
     const dbService = new DatabaseService(MongoDB.client, collection, fieldList)
-
+    if (checkUsed("all")) {
+      return res.send(
+        ResponseTemplate(
+          false,
+          "Không thể xóa! Nhà xuất bản đang được sử dụng trong hệ thống!",
+          null
+        )
+      )
+    }
     let deletedCount = await dbService.deleteAll()
     return res.send(
       ResponseTemplate(true, `${deletedCount} ${collection} deleted`, null)
@@ -170,4 +200,23 @@ exports.deleteAll = async (req, res, next) => {
       new ApiError(500, `An Error occurred while removing all ${collection}`)
     )
   }
+}
+async function checkUsed(id) {
+  //check being used
+  const sachDbService = new DatabaseService(
+    MongoDB.client,
+    sachCollection,
+    sachFieldList
+  )
+  if (id == "all") {
+    let checkUsed = await sachDbService.find({})
+    if (checkUsed.length > 0) return true
+  }
+  if (id != "all") {
+    let checkUsed = await sachDbService.find({ MaNXB: id })
+    if (checkUsed.length > 0) return true
+  }
+
+  return false
+  //
 }
